@@ -1,59 +1,40 @@
 {
-  "name": "ngcomputervision",
+  "name": "client-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:ssr": "ng run ngComputerVision:server:dev",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "11.2.3",
-    "@angular/common": "11.2.3",
-    "@angular/compiler": "11.2.3",
-    "@angular/core": "11.2.3",
-    "@angular/forms": "11.2.3",
-    "@angular/platform-browser": "11.2.3",
-    "@angular/platform-browser-dynamic": "11.2.3",
-    "@angular/platform-server": "11.2.3",
-    "@angular/router": "11.2.3",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.3.1",
-    "core-js": "^3.3.3",
-    "jquery": "3.4.1",
-    "oidc-client": "^1.9.1",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "~12.1.0",
+    "@angular/common": "~12.1.0",
+    "@angular/compiler": "~12.1.0",
+    "@angular/core": "~12.1.0",
+    "@angular/forms": "~12.1.0",
+    "@angular/platform-browser": "~12.1.0",
+    "@angular/platform-browser-dynamic": "~12.1.0",
+    "@angular/router": "~12.1.0",
+    "bootstrap": "^5.1.3",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.2.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.1",
-    "@angular/cli": "11.2.1",
-    "@angular/compiler-cli": "11.2.3",
-    "@angular/language-service": "11.2.3",
+    "@angular-devkit/build-angular": "~12.1.0",
+    "@angular/cli": "~12.1.0",
+    "@angular/compiler-cli": "~12.1.0",
     "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.1.1",
+    "jasmine-core": "~3.7.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "4.0.7"
-  },
-  "optionalDependencies": {
-    "node-sass": "^4.12.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.4.1",
-    "tslint": "~6.1.0"
+    "typescript": "~4.3.2"
   }
 }
