@@ -1,4 +1,9 @@
 export class AvailableLanguage {
-    languageID: string;
-    languageName: string;
+  languageID: string;
+  languageName: string;
+
+  constructor() {
+    this.languageID = '';
+    this.languageName = '';
+  }
 }
