@@ -1,4 +1,9 @@
 export class OcrResult {
-    language: string;
-    detectedText: string
+  language: string;
+  detectedText: string
+
+  constructor() {
+    this.language = '';
+    this.detectedText = '';
+  }
 }
