@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { ComputervisionService } from './computervision.service';
 
 describe('ComputervisionService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: ComputervisionService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ComputervisionService);
+  });
 
   it('should be created', () => {
-    const service: ComputervisionService = TestBed.get(ComputervisionService);
     expect(service).toBeTruthy();
   });
 });
